requirements.txt
asgiref==3.9.1
Django==5.1.4
djangorestframework==3.15.2
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
typing-extensions>=4.6
xhtml2pdf==0.2.15
reportlab==4.0.9
Pillow==10.4.0